+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 
-## 2.0.0 (May 11nd, 2020)
+## 2.1.1 (May 13th, 2020)
+
+**Implemented enhancements:**
+
+- `autoResetKey` is a new prop that allows resetting the animation when the key changes. It works similar to React's `key` prop
+
+## 2.0.1 (May 12th, 2020)
+
+**Bug fix:**
+
+Fix an issue where resetting the animation once it was done when the `isPlaying` was set to true did not start playing after the reset
+
+## 2.0.0 (May 11th, 2020)
 
 **Breaking Changes:**
 
@@ -8,7 +20,7 @@
 - `durationSeconds` is renamed to just `duration`. The duration is now set in **seconds**
 - `startAt` now expects value in **seconds**
 - `onComplete` will receive as an argument the `totalElapsedTime` in **seconds**. The animation can now be repeated by returning an object instead of an array. The object may have the following params: `shouldRepeat` indicates if the loop should start over; `delay` - delay before looping again in seconds; `newStartAt` set new start at value. |
-- Changing the `duration` while the loop is running will update the duration - if the new duration is more than the previous one, the measurement of the elapsed time will continue to the new duration; if the duration is less than the previous one, the `onComplete` callback will be fired. This behaviour can now be changed by passing `shouldResetOnDurationChange: true` so when a new duration is provided the animation loop will start over.
+- Changing the `duration` while the loop is running will update the duration - if the new duration is more than the previous one, the measurement of the elapsed time will continue to the new duration; if the duration is less than the previous one, the `onComplete` callback will be fired.
 
 **Implemented enhancements:**